GROK-16696: Compute: Sens An: Lookups UPD
Updated lookup tables use in sensitivity analysis
