+++ libraries/compute-utils/function-views/src/sensitivity-analysis-view.ts
@@ -15,8 +15,7 @@ import {combineLatest} from 'rxjs';
 import '../css/sens-analysis.css';
 import {CARD_VIEW_TYPE} from '../../shared-utils/consts';
 import {DOCK_RATIO, ROW_HEIGHT, STARTING_HELP} from './variance-based-analysis/constants';
-
-import {getInputsTable, getLookupsInfo, INPUTS_DF, LOOKUP} from './shared/lookup-tools';
+import {getLookupChoiceInput} from './shared/lookup-tools';
 
 const RUN_NAME_COL_LABEL = 'Run name' as const;
 const supportedInputTypes = [DG.TYPE.INT, DG.TYPE.BIG_INT, DG.TYPE.FLOAT, DG.TYPE.BOOL, DG.TYPE.DATA_FRAME];
@@ -576,76 +575,23 @@ export class SensitivityAnalysisView {
     this.runIcon.hidden = this.runButton.disabled;
   }
 
-  private async getLookupChoiceInput(inputsLookup?: string) {
+  private async getLookupElement(inputsLookup?: string) {
     if (inputsLookup === undefined)
       return null;
 
-    const info = getLookupsInfo(inputsLookup);
+    const constIputs = new Map<string, DG.InputBase>();
+    Object.keys(this.store.inputs).forEach((name) => constIputs.set(name, this.store.inputs[name].constForm[0]));
 
-    if ((info === null) || (info.choices === undefined))
-      return null;
+    const lookupElement = await getLookupChoiceInput(inputsLookup, constIputs);
 
-    const inputsDf = await getInputsTable(info.choices);
+    if (lookupElement !== null)
+      lookupElement.input.root.insertBefore(getSwitchMock(), lookupElement.input.captionLabel);
 
-    if (inputsDf === null)
-      return null;
-
-    const cols = inputsDf.columns;
-    const rowCount = inputsDf.rowCount;
-
-    const inpSetsNames = cols.byIndex(INPUTS_DF.INPUT_SETS_COL_IDX).toList();
-    const choices = [LOOKUP.DEFAULT as string].concat(inpSetsNames);
-
-    const defaultInputs = new Map<string, any>();
-    const inputNames = Object.keys(this.store.inputs);
-
-    inputNames.forEach((name) => {
-      defaultInputs.set(name, this.store.inputs[name].const.value);
-    });
-
-    const tableInputs = new Map<string, Map<string, number>>(); // set <-> {(input <-> value)}
-    const colsRaw = new Map<string, Int32Array | Uint32Array | Float32Array | Float64Array>();
-
-    for (const col of cols) {
-      if (col.isNumerical)
-        colsRaw.set(col.name, col.getRawData());
-    }
-
-    for (let row = 0; row < rowCount; ++row) {
-      const inputs = new Map<string, number>();
-      colsRaw.forEach((arr, name) => inputs.set(name, arr[row]));
-      tableInputs.set(inpSetsNames[row], inputs);
-    }
-
-    // create input for lookup table use
-    const lookupChoiceInput = ui.input.choice<string>(info.caption, {
-      items: choices,
-      nullable: false,
-      value: choices[0],
-      tooltipText: info.tooltip,
-      onValueChanged: (value) => {
-        if (value === LOOKUP.DEFAULT)
-          inputNames.forEach((name) => this.store.inputs[name].constForm[0].value = defaultInputs.get(name));
-        else {
-          const colInputs = tableInputs.get(value);
-          inputNames.forEach((name) => {
-            const input = this.store.inputs[name].constForm[0]
-            input.value = colInputs!.get(name) ?? input.value;
-          });
-        }
-      },
-    });
-
-    lookupChoiceInput.root.insertBefore(getSwitchMock(), lookupChoiceInput.captionLabel);
-    
-    return {
-      input: lookupChoiceInput,
-      category: info.category ?? 'Misc',
-    };
+    return lookupElement;
   }
 
   private async buildFormWithBtn(inputsLookup?: string) {
-    const lookupElement = await this.getLookupChoiceInput(inputsLookup);
+    const lookupElement = await this.getLookupElement(inputsLookup);
     
     let prevCategory: string;
